requirements.txt
torch==1.10.0
transformers==4.9.2
datasets==1.11.0
pytesseract==0.3.8
Pillow==8.4.0
streamlit==0.88.0
tqdm==4.62.3
pandas==1.3.3
